samples/apps/cap/py/devgptcap/requirements.txt
pyzmq
protobuf
termcolor
devgpt
